Change course detail look
On some flags it looks a bit strange when we write the name of the
language over the flag. It looks like the flag is smaller than it should
be because the alpha of 0.9 looks sometimes so dense that I didn't
realize that the flag continues behind it.

I experimented a bit with the look and now changed it so the language is
displayed above the flag.