Restore previous main colors
Some of the color changes have been very nice, the primary ones however
I don't like so much.
I have to admit that I did not come up with a nice colorscheme myself
yet, but I would like to propose to restore the old blueish colors.
Orange seems to be a bit too close to red, which just signals "something
is wrong" to me :-)