Handle audio focus loss
Previously, the app crashed when it lost audio focus because another app
requested the focus. This was because mediaPlayer was set to null and
was not initialized again when regaining focus and resuming playback.
Now mediaPlayer is being initialized in loadSentences if it is null.
A few additional changes were necessary, since now the app started
playing automatically when another app had the audio focus and the power
button was pressed to activate the lock screen. This leads to onResume()
being called, which previously started the playback automatically by
calling startSession() in connectToService(). To prevent this, I removed
startSession() from connectToService() and instead call it separately in
onCreateView().